+++ web/packages/shared/storage/localStorage/helpers.ts
@@ -5,8 +5,6 @@ export const getToken = (): string => {
     return token;
 };
 
-export const getUserID = () => getData(LS_KEYS.USER)?.id;
-
 export const isFirstLogin = () =>
     getData(LS_KEYS.IS_FIRST_LOGIN)?.status ?? false;
